added authentication for api routes
api routes now have authentication middleware to ensure security

+++ api/app.js
@@ -99,20 +99,21 @@ let verifySession = (req, res, next) => {
 
 /* END MIDDLEWARE  */
 
-
-
-
-/* list routes*/ 
+/* LIST ROUTES */ 
 /**
  * GET /lists
  * Purpose: Get all lists
  */
-app.get('/lists', (req, res) => {
-// We want to return an array of all the lists in the database
+app.get('/lists', authenticate, (req, res) => {
+// We want to return an array of all the lists in the database that belong to the authenticated user
 	
-    List.find({}).then((lists) => {
+    List.find({
+        _userId: req.user_id
+    }).then((lists) => {
         res.send(lists);
-       });
+       }).catch((e) => {
+        res.send(e);
+});
 })
 
 /**
@@ -120,14 +121,16 @@ app.get('/lists', (req, res) => {
  * Purpose: create a new list
  */
 
-app.post('/lists', (req, res) => {
+app.post('/lists', authenticate, (req, res) => {
 // We want to create a new list and return the new list document back to the user (which includes the id)
     // The list information (fields) will be passed in via the JSON request body	
      let title = req.body.title;
 
     let newList = new List({
         title,
+        _userId: req.user_id
           });
+
     newList.save().then((listDoc) => {
         // the full list document is returned (incl. id)
         res.send(listDoc);
@@ -138,9 +141,9 @@ app.post('/lists', (req, res) => {
  * PATCH /lists/:id
  * Purpose: Update a specified list
  */
-app.patch('/lists/:id', (req, res) => {
+app.patch('/lists/:id', authenticate, (req, res) => {
     // We want to update the specified list (list document with id in the URL) with the new values specified in the JSON body of the request
-   List.findOneAndUpdate({ _id: req.params.id }, {
+   List.findOneAndUpdate({ _id: req.params.id, _userId: req.user_id }, {
         $set: req.body
     }).then(() => {
         res.send({ 'message': 'updated successfully'});
@@ -152,12 +155,16 @@ app.patch('/lists/:id', (req, res) => {
  * DELETE /lists/:id
  * Purpose: Delete a list
  */
-app.delete('/lists/:id', (req, res) => {
+app.delete('/lists/:id', authenticate, (req, res) => {
     // We want to delete the specified list (document with id in the URL)
     List.findOneAndRemove({
         _id: req.params.id,
+        _userId: req.user_id
            }).then((removedListDoc) => {
         res.send(removedListDoc);
+
+        // delete all tasks in deleted list
+        deleteTasksFromList(removedListDoc._id);
 
            })
 });
@@ -166,7 +173,7 @@ app.delete('/lists/:id', (req, res) => {
  * GET /lists/:listId/tasks
  * Purpose: Get all tasks in a specific list
  */
-app.get('/lists/:listId/tasks',  (req, res) => {
+app.get('/lists/:listId/tasks', authenticate, (req, res) => {
     // We want to return all tasks that belong to a specific list (specified by listId)
     Task.find({
         _listId: req.params.listId
@@ -174,39 +181,64 @@ app.get('/lists/:listId/tasks',  (req, res) => {
         res.send(tasks);
     })
 });
-app.get('/lists/:listId/tasks',  (req, res) => {
-Task.findOne({
-    _id: req.params.taskId,
-    _listId:  req.params.listId
-}).then((task) => {
-  res.send(task);
-})
-});
 
 /**
  * POST /lists/:listId/tasks
  * Purpose: Create a new task in a specific list
  */
-app.post('/lists/:listId/tasks',  (req, res) => {
+app.post('/lists/:listId/tasks', authenticate, (req, res) => {
     // We want to create a new task in a list specified by listId
 
-                let newTask = new Task({
+    List.findOne({
+        _id: req.params.listId,
+        _userId: req.user_id
+    }).then((list) => {
+        if(list) {
+            // list object with spec. conditions valid
+            // currently authenticated user can create new tasks
+            return true;
+        }
+
+        return false;
+    }).then((canCreateTask) => {
+        if(canCreateTask) {
+            let newTask = new Task({
                 title: req.body.title,
                 _listId: req.params.listId
             });
             newTask.save().then((newTaskDoc) => {
                 res.send(newTaskDoc);
             })
+        }
+        else {
+            res.sendStatus(404);
+        }
+    })
+
+                
 })
 
 /**
  * PATCH /lists/:listId/tasks/:taskId
  * Purpose: Update an existing task
  */
-app.patch('/lists/:listId/tasks/:taskId', (req, res) => {
+app.patch('/lists/:listId/tasks/:taskId', authenticate, (req, res) => {
     // We want to update an existing task (specified by taskId)
 
-                Task.findOneAndUpdate({
+    List.findOne({
+        _id: req.params.listId,
+        _userId: req.user_id
+    }).then((list) => {
+        if(list) {
+            // list object with spec. conditions valid
+            // currently authenticated user can update tasks within list
+            return true;
+        }
+
+        return false;
+    }).then((canUpdateTasks) => {
+        if (canUpdateTasks){
+            Task.findOneAndUpdate({
                 _id: req.params.taskId,
                 _listId: req.params.listId
             }, {
@@ -215,24 +247,46 @@ app.patch('/lists/:listId/tasks/:taskId', (req, res) => {
             ).then(() => {
                 res.send({ message: 'Updated successfully.' })
             })
-          });
+        } else {
+            res.sendStatus(404);
+        }
+    })
 
+                
+          });
 
 /**
  * DELETE /lists/:listId/tasks/:taskId
  * Purpose: Delete a task
  */
-app.delete('/lists/:listId/tasks/:taskId', (req, res) => {
+app.delete('/lists/:listId/tasks/:taskId', authenticate, (req, res) => {
 
-               Task.findOneAndRemove({
+    List.findOne({
+        _id: req.params.listId,
+        _userId: req.user_id
+    }).then((list) => {
+        if(list) {
+            // list object with spec. conditions valid
+            // currently authenticated user can update tasks within list
+            return true;
+        }
+
+        return false;
+    }).then((canUpdateTasks) => {
+        if (canUpdateTasks) {
+            Task.findOneAndRemove({
                 _id: req.params.taskId,
                 _listId: req.params.listId
             }).then((removedTaskDoc) => {
                 res.send(removedTaskDoc);
             })
         
-    });
+    } else {
+        res.sendStatus(404);
+    }})
+        }
 
+    );
     
     /* USER ROUTES */
 
@@ -309,7 +363,16 @@ app.get('/users/me/access-token', verifySession, (req, res) => {
     });
 })
 
+/* HELPER METHODS */
+
+let deleteTasksFromList = (_listId) => {
+    Task.deleteMany({
+        _listId
+    }).then(() => {
+        console.log("Tasks from " + _listId + "were deleted.");
+    });
+}
+
 app.listen(3000, () => {
     console.log("Server is listening on port 3000");
 })
-
